feat(module-api): Fixed wrong package hierarchy
