--- server/movieflex_Server/src/main/java/io/egen/Entity/movie_field.java
@@ -1,143 +1,0 @@
-package io.egen.Entity;
-
-public class movie_field {
-	
-	private String Title;
-    private int Year;
-    private String Rated;
-    private String Released;
-    private String Runtime;
-    private String Genre;
-    private String Director;
-    private String Writer;
-    private String Actors;
-    private String Plot;
-    private String Language;
-    private String Country;
-    private String Awards;
-    private String Poster;
-    private int Metascore;
-    private float imdbRating;
-    private int imdbVotes;
-    private String imdbID;
-    private String Type;
-    
-    
-	public String getTitle() {
-		return Title;
-	}
-	public void setTitle(String title) {
-		Title = title;
-	}
-	public int getYear() {
-		return Year;
-	}
-	public void setYear(int year) {
-		Year = year;
-	}
-	public String getRated() {
-		return Rated;
-	}
-	public void setRated(String rated) {
-		Rated = rated;
-	}
-	public String getReleased() {
-		return Released;
-	}
-	public void setReleased(String released) {
-		Released = released;
-	}
-	public String getRuntime() {
-		return Runtime;
-	}
-	public void setRuntime(String runtime) {
-		Runtime = runtime;
-	}
-	public String getGenre() {
-		return Genre;
-	}
-	public void setGenre(String genre) {
-		Genre = genre;
-	}
-	public String getDirector() {
-		return Director;
-	}
-	public void setDirector(String director) {
-		Director = director;
-	}
-	public String getWriter() {
-		return Writer;
-	}
-	public void setWriter(String writer) {
-		Writer = writer;
-	}
-	public String getActors() {
-		return Actors;
-	}
-	public void setActors(String actors) {
-		Actors = actors;
-	}
-	public String getPlot() {
-		return Plot;
-	}
-	public void setPlot(String plot) {
-		Plot = plot;
-	}
-	public String getLanguage() {
-		return Language;
-	}
-	public void setLanguage(String language) {
-		Language = language;
-	}
-	public String getCountry() {
-		return Country;
-	}
-	public void setCountry(String country) {
-		Country = country;
-	}
-	public String getAwards() {
-		return Awards;
-	}
-	public void setAwards(String awards) {
-		Awards = awards;
-	}
-	public String getPoster() {
-		return Poster;
-	}
-	public void setPoster(String poster) {
-		Poster = poster;
-	}
-	public int getMetascore() {
-		return Metascore;
-	}
-	public void setMetascore(int metascore) {
-		Metascore = metascore;
-	}
-	public float getImdbRating() {
-		return imdbRating;
-	}
-	public void setImdbRating(float imdbRating) {
-		this.imdbRating = imdbRating;
-	}
-	public int getImdbVotes() {
-		return imdbVotes;
-	}
-	public void setImdbVotes(int imdbVotes) {
-		this.imdbVotes = imdbVotes;
-	}
-	public String getImdbID() {
-		return imdbID;
-	}
-	public void setImdbID(String imdbID) {
-		this.imdbID = imdbID;
-	}
-	public String getType() {
-		return Type;
-	}
-	public void setType(String type) {
-		Type = type;
-	}
-    
-    
-
-}
